==> 00_marketdata.ipynb <==
import requests
import time
import pandas as pd
import os

# Get list of djia tickers
dow = ['mmm', 'axp', 'aapl', 'ba', 'cat', 'cvx', 'csco', 'ko', 'dow', 'xom', 'gs', 'hd', 'ibm',
      'intc', 'jnj', 'jpm', 'mcd', 'mrk', 'msft', 'nke', 'pfe', 'pg', 'rtx', 'trv', 'unh',
      'vz', 'v', 'wmt', 'wba', 'dis']


key = 'XM9DAWJLXUGZAO5A'
base_url = "https://www.alphavantage.co/query" 
sym = ''


sym = 'BA'
res = requests.get(base_url,
                  params={
    'function' : 'TIME_SERIES_INTRADAY',
    'symbol' : sym,
    'interval' : '1min',
    'outputsize' : 'full',
    'apikey' : key
})
res

for sym in dow[:2]:
    successful = False
    while not successful:
        res = requests.get(base_url,
                      params={
        'function' : 'TIME_SERIES_INTRADAY',
        'symbol' : sym.upper(),
        'interval' : '1min',
        'outputsize' : 'full',
        'apikey' : key})
        if res.status_code == 200:
            print(f'Successfully pulled {sym} data')
            successful = True
        else:
            print(f'trying to get {sym} again')
            print(res)
            time.sleep(5)
    stock = res.json()
    df = pd.DataFrame(stock['Time Series (1min)']).T
    df.to_csv(f'./datasets/dow/{sym}.csv')
    print(f'output {sym} to csv')
    time.sleep(30)
    

aapl = res.json()b

df = pd.DataFrame(aapl['Time Series (1min)']).T
df.to_csv('./asdfasdf.csv')

pd.read_csv('./asdfasdf.csv').head()

df.tail()

res.status_code

df2 = pd.read_csv('./datasets/russia site/aapl.csv')

df2.drop(columns=['<PER>','<TICKER>']).head()

df2['<PER>'].value_counts()

df2.sort_values('<DATE>', inplace=True)

df2.describe()

https://www.finam.ru/profile/akcii-usa-bats/united-technologies/export/?market=25&em=18134&token=03AGdBq26R5oYsaVJ7QZMxHCfVrppyq7nlA71Irgcsrpob_UH7svIif7RzoRXw-qYD0ePNpdUvBpjv9DPKuJ4oMCyZ8d72rlRSXlZJxEUw62eVdvK1IPAg-8gEapI_g6Ou3xpdrs-497L7mGmJnplVhaOd7_ZxpVrTXFB_X-g5zpwHkz_2X49qfVMz3UqpcuXbVcnDiXeBoHWKMHxHUFIkGh-TdUvewDnvFIZgMJ7K1EsXjJCA-EOUHmvCYrpqrG03CMRXnhhVXwnaiSF1MRqSaZugbaUPzL6PgowRw8aY8NwVRkK4toG2vrnxMP_SigIfhdzz5JH17DITkvC17Lpxxjc9xhj4JuLnD28MSsus8LbGcsSyG68RlO61tkNMcI2kQTL0Odql1FIh8CEJpGE6uc9EsxMh0wXiBOQANDkgbAVQISTQQ-veXdGd_7VpZBdfVFgx-j2epsGh&code=UTX&apply=0&df=3&mf=8&yf=2019&from=03.09.2019&dt=31&mt=11&yt=2019&to=31.12.2019&p=1&f=UTX_190903_191231&e=.txt&cn=UTX&dtf=1&tmf=1&MSOR=1&mstimever=0&sep=1&sep2=1&datf=6&at=1

def clean_ru_data(df):
    new_col_names= {
        '<TICKER>' : 'ticker',
        '<DATE>' : 'date',
        '<TIME>' : 'time',
        '<LAST>' : 'price',
        '<VOL>' : 'vol'
    }
    df.rename(columns=new_col_names, inplace=True)
    df['ticker'] = df['ticker'].str.lower()
    df['ticker'] = df['ticker'].str.replace('us1.','')
    df['date_time'] = df['date'].astype(str) + df['time'].astype(str)
    df['date_time'] = pd.to_datetime(df['date_time'], format='%Y%m%d%H%M%S')
    df = df.drop(columns='<PER>')
    return df

files = os.listdir('./datasets/russia site/dirty')
count = 0
for file in files:
    df = pd.read_csv(f'./datasets/russia site/dirty/{file}')
    df = clean_ru_data(df)
    df.to_csv(f'./datasets/russia site/cleaned/{file}', index=False)
    count += 1
    print(f'cleaned and output: {file}.  {count}/31')

files = os.listdir('./datasets/russia site/cleaned')
count = 0
for file in files:
    df = pd.read_csv(f'./datasets/russia site/cleaned/{file}')
    print(f'{file} has {df.shape[0]} rows')

pd.read_csv('/Users/jsf/Downloads/US1.RTX_190903_200411.txt').tail()

df3 = clean_ru_data(df2)
df3.head()

df2

# df3['date_'] = pd.to_datetime(df3['date'], format='%Y%m%d')
# df3['time_'] = pd.to_datetime(df3['time'], format='%H%M%S')
df3['date_time_'] = pd.to_datetime(df3['date_time'], format='%Y%m%d%H%M%S')
df3.head()

# 

# talk to twitter again for api access for searching for tweets using stock tickers with a $
# maybe other twitter scrapers can work or do it via tweetdeck/other app
