==> 01_twitterdata.ipynb <==
import pandas as pd
import requests
import GetOldTweets3 as got
import time
import datetime
import os
from ratelimit import limits, sleep_and_retry

# Get list of djia tickers
dow = ['mmm', 'axp', 'aapl', 'ba', 'cat', 'cvx', 'csco', 'ko', 'dow', 'xom', 'gs', 'hd', 'ibm',
      'intc', 'jnj', 'jpm', 'mcd', 'mrk', 'msft', 'nke', 'pfe', 'pg', 'trv', 'unh',
      'vz', 'v', 'wmt', 'wba', 'dis', 'rtn', 'utx'] #  
query_dow = [f'${i}' for i in dow]


def get_twitter_info(tweet_df):
    tweet_df["id"] = tweet_df["got_criteria"].apply(lambda x: x.id)
    tweet_df["tweet_text"] = tweet_df["got_criteria"].apply(lambda x: x.text)
    tweet_df["date"] = tweet_df["got_criteria"].apply(lambda x: x.date)
    tweet_df["hashtags"] = tweet_df["got_criteria"].apply(lambda x: x.hashtags)
    tweet_df["username"] = tweet_df["got_criteria"].apply(lambda x: x.username)
    tweet_df["mentions"] = tweet_df["got_criteria"].apply(lambda x: x.mentions)
    tweet_df["rewtweets"] = tweet_df["got_criteria"].apply(lambda x: x.retweets)
    tweet_df["replies"] = tweet_df["got_criteria"].apply(lambda x: x.replies)
    tweet_df["link"] = tweet_df["got_criteria"].apply(lambda x: x.permalink)
    return tweet_df
# https://medium.com/@robbiegeoghegan/download-twitter-data-with-10-lines-of-code-42eb2ba1ab0f
#set search variables

oldest_date = datetime.datetime.strptime("2019-09-03", "%Y-%m-%d")
newest_date = datetime.datetime.strptime("2019-12-31", "%Y-%m-%d")
# add in iterator for week or date
date_range = [oldest_date + datetime.timedelta(days = x) for x in range(0,(newest_date-oldest_date).days, 1)]
# https://www.pythonprogramming.in/get-range-of-dates-between-specified-start-and-end-date.html

# tweetCriteria_list = []
# for keyword in query_dow:
#     tweetCriteria = got.manager.TweetCriteria().setQuerySearch(keyword)\
#                                             .setSince(oldest_date)\
#                                             .setUntil(newest_date)
#     tweetCriteria_list.append(tweetCriteria)
#create twitter info for each company



# @sleep_and_retry
# @limits(calls=15, period=60)
def pull_tweets(company):
    
    oldest_date = datetime.datetime.strptime("2019-09-03", "%Y-%m-%d")
    newest_date = datetime.datetime.strptime("2019-12-31", "%Y-%m-%d")
    # add in iterator for week or date
    date_range = [oldest_date + datetime.timedelta(days = x) for x in range(0,(newest_date-oldest_date).days, 1)]
#     for company in dow:
    tweet_list = []
    tweet_dict = {}
    for i, week in enumerate(date_range[:-1]):
        start = str(date_range[i])[:10]
        end = str(date_range[i+1])[:10]
        criteria = got.manager.TweetCriteria().setQuerySearch(f'${company}')\
                                            .setSince(start)\
                                            .setUntil(end)
#                                             .setMaxTweets(3_000)
            
        tweets = got.manager.TweetManager.getTweets(criteria)
        tweet_list += tweets
        if (i+1) % 7 == 0:
            print(f'added week ended {end} for {company}.  {i/len(date_range):.2%}')
            print(f'currently at {len(tweet_list)} tweets')
#                 time.sleep(60)

        
    print(f'pulled {len(tweet_list)} tweets for {company}.')
    #create df
    tweet_dict[company] = tweet_list
    tweet_df = pd.DataFrame(dict([ (k,pd.Series(v)) for k,v in tweet_dict.items() ]))
    tweet_df['tweet_count'] = tweet_df.index
    tweet_df = pd.melt(tweet_df, id_vars=["tweet_count"], var_name='company', value_name='got_criteria')
    tweet_df = get_twitter_info(tweet_df)
    tweet_df = tweet_df.drop("got_criteria", 1)
    tweet_df.to_csv(f'./datasets/twitters/{company}_tweets.csv', index = False)
#         time.sleep(180)


# tweet_df = tweet_df.dropna()
#extract twitter information



count = 0
dow = ['dis', 'rtn', 'utx'] 
# 'mmm', 'axp', 'aapl', 'ba', 'cat', 'cvx', 'csco', 'ko', 'dow', 'xom', 'gs', 'hd', 'ibm',
#       'intc', 'jnj', 'jpm', 'mcd', 'mrk', 'msft', 'nke', 'pfe', 
# 'pg', 'trv', 'unh', 'vz', 'v', 'wmt', 'wba', 
for company in dow:
    pull_tweets(company)
    count += 1
    print(f'pulled {count}/{len(dow)}')
    

for file in os.listdir('./datasets/twitters/'):
    df = pd.read_csv(f'./datasets/twitters/{file}')
    print(f'{file[:-4]} is {df.shape[0]} long')

os.listdir('./datasets/twitters/')

https://github.com/Mottl/GetOldTweets3/issues/57
https://github.com/Mottl/GetOldTweets3/issues/3#issuecomment-527642499
https://github.com/libbyh/GetOldTweets-python/blob/py3/GetOldTweets3/manager/TweetManager.py#L7
https://github.com/libbyh/GetOldTweets-python/blob/py3/GetOldTweets3/manager/TweetManager.py
https://github.com/tomasbasham/ratelimit
https://github.com/Mottl/GetOldTweets3/blob/master/bin/GetOldTweets3
