==> 06_twitterdataprep.ipynb <==
import numpy as np
import pandas as pd
import os
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer


directory = './datasets/twitters/'
files = [file for file in os.listdir(directory) if file.endswith('.csv')]

counter = 0
stocks = {}
for file in files:
    df = pd.read_csv(directory + file)
    df['date'] = pd.to_datetime(df['date']).dt.tz_convert('US/Eastern')
    df.set_index('date', inplace=True)
    df.index = pd.DatetimeIndex(df.index).to_period('ms')
    df.sort_index(inplace=True)
    
    sent = SentimentIntensityAnalyzer()
    df['vader_sentiment'] = df['tweet_text'].apply(lambda x: sent.polarity_scores(str(x))['compound'])
    
    stocks[file[:-11]] = df
    df.to_csv(f'./datasets/twitters/withsent/{df["company"][0]}.csv', index=True)
    counter+=1
    print(f'exported {df["company"][0]}.  {counter}/{len(files)}')
    

df = pd.read_csv('./datasets/twitters/wba_tweets.csv')
df.head()

df['vader_sentiment'] = df['tweet_text'].apply(lambda x: sent.polarity_scores(x)['compound'])

df.head()

for i in range(10):
    print(df['tweet_text'][i])
    print(df['vader_sentiment'][i])

sent = SentimentIntensityAnalyzer()

str_list = [
    "The train to London leaves at four o'clock",
    'This blog post is good.',
    'This blog post is good. It was very helpful. The author is amazing.',
    'This blog post is very good. This author is a horrible writer usually, but here he got lucky.'
]
for x in str_list:
    print(sent.polarity_scores(x))

df['date'] = pd.to_datetime(df['date']).dt.tz_convert('US/Eastern')
df.set_index('date', inplace=True)
df.index = pd.DatetimeIndex(df.index).to_period('ms')
df.sort_index(inplace=True)
df.head()

df2 = df['vader_sentiment'].resample('s').sum()

df2.head()

# from google.cloud import language
# from google.cloud.language import enums, types


# def analyze_text_sentiment(text):
#     client = language.LanguageServiceClient()
#     document = types.Document(
#         content=text,
#         type=enums.Document.Type.PLAIN_TEXT)

#     response = client.analyze_sentiment(document=document)

#     sentiment = response.document_sentiment
#     results = [
#         ('text', text),
#         ('score', sentiment.score),
#         ('magnitude', sentiment.magnitude),
#     ]
#     for k, v in results:
#         print('{:10}: {}'.format(k, v))

# analyze_text_sentiment(str_list[1])
