==> 08_addingsent.ipynb <==
import numpy as np
import pandas as pd
import os


comp = pd.read_csv('./working.csv', index_col='date_time', parse_dates=True)
comp.sort_index(inplace=True)

twit = pd.read_csv('./datasets/twitters/withsent/csco.csv', index_col='date', parse_dates=True)
twit.sort_index(inplace=True)


twit_score = twit[['vader_sentiment']]
twit_score

twit_score = twit[['vader_sentiment']]
twit_score

twit_score.resample('s').mean()

twit_score['vader_sentiment'].ewm(50).mean().isna().sum()

twit_score['mean_sent'] = twit_score['vader_sentiment'].ewm(50).mean()
twit_score.head()

df2 = pd.merge_asof(comp, twit_score, left_index=True, right_index=True, allow_exact_matches=True, direction='nearest', 
                    tolerance=pd.Timedelta('12H'))
# df2['vader_sentiment'].fillna(0, inpalace=True)


df2[['vader_sentiment']]

twit_score.index.dtype

df3 = pd.merge_ordered(comp, twit_score, left_on:comp.index, right_on:twit_score.index)

comp.index.dtype

twit_score.between_time('16:00', '9:30')

df = pd.merge(comp, twit_score, left_index=True, right_index=True, how='left')
df['vader_sentiment'].fillna(0, inplace=True)
df

df[df['vader_sentiment']!=0]

twit_score[twit_score.index.duplicated()]

(comp['vol'] == 0).sum()

twit.head()


twit_score

rs = twit_score.resample('min').mean()

rs.ewm(50).mean()
